% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compas}
\alias{compas}
\title{Compas}
\format{
\subsection{\code{compas}}{

A data frame with 6172 rows and 9 columns:
}
}
\source{
\url{https://mlr3fairness.mlr-org.com/reference/compas.html}
}
\usage{
compas
}
\description{
Recidivism data popular in fairness analysis
}
\keyword{datasets}
